Report 2
This paper was based on a csv file which included the data from 2015 to the 2017 season, and consists of tables describing constructors, race drivers, lap times, pit stops and more. The main tasks were to visualize the data, and then hunt for the changing trend of race speed, and explore the relations between race drivers. constructors, the accumulated points of both and race speed.